refactor Synchronizer::Objets::Revision into a module
